Add new genetic algorithm for logical expressions
Added a new class GeneticAlgorithm which uses tournament selection with
reproduction and mutation. Mutation is done using some methods that I
added to the Gene, Conjugate, Literal and Value classes. The mutation is
currently quite drastic, sometimes replacing the entire expression. This
can be mitigated by insisting that only Literals or Values are mutated,
or that if an expression is made up of only one Conjugate, only mutate
the Literals. Performance is a bit hit or miss, managed to get fitness
of 180 in one run and then 91 in another. Doesn't integrate the weight
optimisation yet. Also removed old files since we're not using them
anymore.