Add crossover
Added crossover to the genetic algorithm. It mostly works, the deepest
it can exchange at the moment is Literal.value i.i. it can't exchange
weights. I added two new methods to each expression class, get_node and
replace. get_node goes to a random depth in an expression and returns a
sub-expression from that depth as well as a list of indices that can be
used to get back to where the sub-expression was stored. replace then
takes the sub-expression that was retrieved from the other gene and
follows the path list to replace the original sub-expression with the
new one.