Move c constant to Value class, numerous bug fixes
Moved the c constant from the Literal class to the Value class in order
to include said constant in the exchange during a crossover event.
Updated the update_weights method accordingly in the Gene class. Fixed a
bug in the mutation method of the Value class in which a the same values
were being generated rather than new ones during a mutation.